bug fixed in sum calculation

+++ Hm-3/calculator.js
@@ -5,7 +5,7 @@
 function Calculator() {
 
 	this.sum = function(number1, number2) {
-		return parseInt(number1) + parseInt(number2);
+		return parseFloat(number1) + parseFloat(number2);
 	}
 	this.subtraction = function(number1, number2) {
 		return number1 - number2;
